+++ changelog.md
@@ -8,6 +8,7 @@ vX.X.X
 - Fixed bug where build configurations were not being included in repo, closes #85.
 - Added logging sub-tab and basic logging functionality, closes #96.
 - Greyed out textfield and browse button on logging tab when logging is active, closes #97.
+- Added ability to select between "overwrite" and "append" logging methods, closes #99.
 
 v0.4.1
 ------